+++ cut-core.js
@@ -5,7 +5,7 @@
  * @license
  */
 
-DEBUG = (typeof DEBUG === 'undefined' || DEBUG) && console;
+DEBUG = typeof DEBUG === 'undefined' || DEBUG;
 
 function Cut() {
   if (!(this instanceof Cut)) {
@@ -34,7 +34,11 @@ function Cut() {
   this._tickAfter = [];
 
   this._alpha = 1;
-};
+
+  this._attrs = null;
+  this._listeners = null;
+  this._flags = null;
+}
 
 Cut._create = (function() {
   if (typeof Object.create == 'function') {
@@ -56,6 +60,8 @@ Cut._create = (function() {
     };
   }
 })();
+
+Cut._TS = 0;
 
 Cut._stats = {
   create : 0,
@@ -167,7 +173,7 @@ Cut.prototype.toString = function() {
 };
 
 Cut.prototype.id = function(id) {
-  if (!arguments.length) {
+  if (typeof id === 'undefined') {
     return this._id;
   }
   this._id = id;
@@ -175,28 +181,56 @@ Cut.prototype.id = function(id) {
 };
 
 Cut.prototype.attr = function(name, value) {
-  if (arguments.length < 2) {
-    return this._attrs ? this._attrs[name] : undefined;
+  if (typeof value === 'undefined') {
+    return this._attrs !== null ? this._attrs[name] : undefined;
   }
-  (this._attrs ? this._attrs : this._attrs = {})[name] = value;
+  (this._attrs !== null ? this._attrs : this._attrs = {})[name] = value;
   return this;
 };
 
 Cut.prototype.on = Cut.prototype.listen = function(types, listener) {
-  if (typeof listener !== 'function') {
-    return this;
-  }
-  types = (Cut._isArray(types) ? types.join(' ') : types).split(/\s+/);
-  for (var i = 0; i < types.length; i++) {
-    var type = types[i];
-    if (type) {
-      this._listeners = this._listeners || {};
+  if (types = this._onoff(types, listener, true)) {
+    for (var i = 0; i < types.length; i++) {
+      var type = types[i];
       this._listeners[type] = this._listeners[type] || [];
       this._listeners[type].push(listener);
-      this._listenOn(type);
+      this._flag(type, true);
     }
   }
   return this;
+};
+
+Cut.prototype.off = function(types, listener) {
+  if (types = this._onoff(types, listener, false)) {
+    for (var i = 0; i < types.length; i++) {
+      var type = types[i], all = this._listeners[type], index;
+      if (all && (index = all.indexOf(listener)) >= 0) {
+        all.splice(index, 1);
+        if (!all.length) {
+          delete this._listeners[type];
+        }
+        this._flag(type, false);
+      }
+    }
+  }
+  return this;
+};
+
+Cut.prototype._onoff = function(types, listener, create) {
+  if (!types || !types.length || typeof listener !== 'function') {
+    return false;
+  }
+  if (!(types = (Cut._isArray(types) ? types.join(' ') : types).match(/\S+/g))) {
+    return false;
+  }
+  if (this._listeners === null) {
+    if (create) {
+      this._listeners = {};
+    } else {
+      return false;
+    }
+  }
+  return types;
 };
 
 Cut.prototype.listeners = function(type) {
@@ -219,24 +253,24 @@ Cut.prototype.trigger = function(name, args) {
   return this;
 };
 
-Cut.prototype.visit = function(visitor) {
+Cut.prototype.visit = function(visitor, data) {
   var reverse = visitor.reverse;
   var visible = visitor.visible;
-  if (visitor.start && visitor.start(this)) {
+  if (visitor.start && visitor.start(this, data)) {
     return;
   }
   var child, next = reverse ? this.last(visible) : this.first(visible);
   while (child = next) {
     next = reverse ? child.prev(visible) : child.next(visible);
-    if (child.visit(visitor, reverse)) {
+    if (child.visit(visitor, data)) {
       return true;
     }
   }
-  return visitor.end && visitor.end(this);
+  return visitor.end && visitor.end(this, data);
 };
 
 Cut.prototype.visible = function(visible) {
-  if (!arguments.length) {
+  if (typeof visible === 'undefined') {
     return this._visible;
   }
 
@@ -333,7 +367,7 @@ Cut.prototype.appendTo = function(parent) {
     parent._first = this;
   }
 
-  this._parent._listenOn(this);
+  this._parent._flag(this, true);
 
   this._ts_parent = Cut._TS++;
   parent._ts_children = Cut._TS++;
@@ -363,7 +397,7 @@ Cut.prototype.prependTo = function(parent) {
     parent._last = this;
   }
 
-  this._parent._listenOn(this);
+  this._parent._flag(this, true);
 
   this._ts_parent = Cut._TS++;
   parent._ts_children = Cut._TS++;
@@ -415,7 +449,7 @@ Cut.prototype.insertBefore = function(next) {
   this._prev = prev;
   this._next = next;
 
-  this._parent._listenOn(this);
+  this._parent._flag(this, true);
 
   this._ts_parent = Cut._TS++;
   this.touch();
@@ -442,7 +476,7 @@ Cut.prototype.insertAfter = function(prev) {
   this._prev = prev;
   this._next = next;
 
-  this._parent._listenOn(this);
+  this._parent._flag(this, true);
 
   this._ts_parent = Cut._TS++;
   this.touch();
@@ -475,7 +509,7 @@ Cut.prototype.remove = function() {
       this._parent._last = this._prev;
     }
 
-    this._parent._listenOff(this);
+    this._parent._flag(this, false);
 
     this._parent._ts_children = Cut._TS++;
     this._parent.touch();
@@ -494,7 +528,7 @@ Cut.prototype.empty = function() {
     next = child._next;
     child._prev = child._next = child._parent = null;
 
-    this._listenOff(child);
+    this._flag(child, false);
   }
 
   this._first = this._last = null;
@@ -504,46 +538,36 @@ Cut.prototype.empty = function() {
   return this;
 };
 
-Cut.prototype._listenOn = function(obj) {
+// Deep flags used for optimizing event distribution.
+Cut.prototype._flag = function(obj, on) {
+  if (typeof on === 'undefined') {
+    return this._flags !== null && this._flags[obj] || 0;
+  }
   if (typeof obj === 'string') {
-    this._listenersCount = this._listenersCount || {};
-    if (!this._listenersCount[obj] && this._parent) {
-      this._parent._listenOn(obj);
+    if (on) {
+      this._flags = this._flags || {};
+      if (!this._flags[obj] && this._parent) {
+        this._parent._flag(obj, true);
+      }
+      this._flags[obj] = (this._flags[obj] || 0) + 1;
+
+    } else if (this._flags && this._flags[obj] > 0) {
+      if (this._flags[obj] == 1 && this._parent) {
+        this._parent._flag(obj, false);
+      }
+      this._flags[obj] = this._flags[obj] - 1;
     }
-    this._listenersCount[obj] = (this._listenersCount[obj] || 0) + 1;
-  } else if (typeof obj === 'object') {
-    if (obj._listenersCount) {
-      for ( var type in obj._listenersCount) {
-        if (obj._listenersCount[type] > 0) {
-          this._listenOn(type);
+  }
+  if (typeof obj === 'object') {
+    if (obj._flags) {
+      for ( var type in obj._flags) {
+        if (obj._flags[type] > 0) {
+          this._flag(type, on);
         }
       }
     }
   }
-};
-
-Cut.prototype._listenOff = function(obj) {
-  if (typeof obj === 'string') {
-    this._listenersCount = this._listenersCount || {};
-    if (this._listenersCount[obj] == 1 && this._parent) {
-      this._parent._listenOff(obj);
-    }
-    this._listenersCount[obj] = Math.max((this._listenersCount[obj] || 0) - 1,
-        0);
-
-  } else if (typeof obj === 'object') {
-    if (obj._listenersCount) {
-      for ( var type in obj._listenersCount) {
-        if (obj._listenersCount[type] > 0) {
-          this._listenOff(type);
-        }
-      }
-    }
-  }
-};
-
-Cut.prototype._listens = function(type) {
-  return (this._listenersCount && this._listenersCount[type]) || 0;
+  return this;
 };
 
 Cut.prototype.touch = function() {
@@ -552,12 +576,21 @@ Cut.prototype.touch = function() {
   return this;
 };
 
-Cut.prototype.pin = function() {
-  if (!arguments.length) {
+Cut.prototype.pin = function(a, b) {
+  if (typeof a === 'object') {
+    this._pin.set(a);
+    return this;
+
+  } else if (typeof a === 'string') {
+    if (typeof b === 'undefined') {
+      return this._pin.get(a);
+    } else {
+      this._pin.set(a, b);
+      return this;
+    }
+  } else if (typeof a === 'undefined') {
     return this._pin;
   }
-  var obj = this._pin.update.apply(this._pin, arguments);
-  return obj === this._pin ? this : obj;
 };
 
 Cut.prototype.matrix = function() {
@@ -640,17 +673,19 @@ Cut.Tween.prototype.tween = function(duration, delay) {
   return this;
 };
 
-Cut.Tween.prototype.pin = function(pin) {
+Cut.Tween.prototype.pin = function(a, b) {
   if (this._next !== this._queue[this._queue.length - 1]) {
     this._owner.touch();
     this._queue.push(this._next);
   }
 
   var end = this._next.end;
-  if (arguments.length === 1) {
-    Cut._extend(end, arguments[0]);
-  } else if (arguments.length === 2) {
-    end[arguments[0]] = arguments[1];
+  if (typeof a === 'object') {
+    for ( var attr in a) {
+      end[attr] = a[attr];
+    }
+  } else if (typeof b !== 'undefined') {
+    end[a] = b;
   }
   return this;
 };
@@ -678,11 +713,10 @@ Cut.root = function(request, render) {
 };
 
 Cut.Root = function(request, render) {
-  Cut.String.prototype._super.apply(this, arguments);
+  Cut.Root.prototype._super.apply(this, arguments);
 
   this._paused = true;
   this._render = render;
-  this._request = request;
 
   var self = this;
   var requestCallback = function() {
@@ -696,7 +730,7 @@ Cut.Root = function(request, render) {
   };
 
   this.request = function() {
-    this._request.call(null, requestCallback);
+    request(requestCallback);
   };
 
   this.on('viewport', function(viewport) {
@@ -762,7 +796,7 @@ Cut.Root.prototype.pause = function() {
 
 Cut.Root.prototype.touch = function() {
   this.resume();
-  return Cut.prototype.touch.apply(this, arguments);
+  return Cut.prototype.touch.call(this);
 };
 
 Cut.Root.prototype.resize = function(width, height, ratio) {
@@ -773,7 +807,7 @@ Cut.Root.prototype.resize = function(width, height, ratio) {
   data.ratio = ratio;
   this.visit({
     start : function(cut) {
-      if (!cut._listens('viewport')) {
+      if (!cut._flag('viewport')) {
         return true;
       }
       cut.publish('viewport', [ data ]);
@@ -784,7 +818,7 @@ Cut.Root.prototype.resize = function(width, height, ratio) {
 
 Cut.image = function(cutout) {
   var image = new Cut.Image();
-  cutout && image.setImage(cutout);
+  cutout && image.image(cutout);
   return image;
 };
 
@@ -796,7 +830,14 @@ Cut.Image.prototype = Cut._create(Cut.prototype);
 Cut.Image.prototype._super = Cut;
 Cut.Image.prototype.constructor = Cut.Image;
 
-Cut.Image.prototype.setImage = function(cutout) {
+/**
+ * @deprecated Use image
+ */
+Cut.Image.prototype.setImage = function(a, b, c) {
+  return this.image(a, b, c);
+};
+
+Cut.Image.prototype.image = function(cutout) {
   this._cutouts[0] = Cut.Out.select(cutout);
   this._cutouts.length = 1;
   this.pin({
@@ -810,11 +851,11 @@ Cut.Image.prototype.setImage = function(cutout) {
 };
 
 Cut.Image.prototype.cropX = function(w, x) {
-  return this.setImage(this._cutouts[0].cropX(w, x));
+  return this.image(this._cutouts[0].cropX(w, x));
 };
 
 Cut.Image.prototype.cropY = function(h, y) {
-  return this.setImage(this._cutouts[0].cropY(h, y));
+  return this.image(this._cutouts[0].cropY(h, y));
 };
 
 Cut.Image.prototype._slice = function(i) {
@@ -994,7 +1035,7 @@ Cut.Image.prototype.stretch = function(inner) {
 };
 
 Cut.anim = function(frames, fps) {
-  var anim = new Cut.Anim().setFrames(frames).gotoFrame(0);
+  var anim = new Cut.Anim().frames(frames).gotoFrame(0);
   fps && anim.fps(fps);
   return anim;
 };
@@ -1035,7 +1076,7 @@ Cut.Anim.prototype.constructor = Cut.Anim;
 Cut.Anim.FPS = 22;
 
 Cut.Anim.prototype.fps = function(fps) {
-  if (!arguments.length) {
+  if (typeof fps === 'undefined') {
     return this._fps;
   }
   this._fps = fps || Cut.Anim.FPS;
@@ -1043,7 +1084,14 @@ Cut.Anim.prototype.fps = function(fps) {
   return this;
 };
 
-Cut.Anim.prototype.setFrames = function(frames) {
+/**
+ * @deprecated Use frames
+ */
+Cut.Anim.prototype.setFrames = function(a, b, c) {
+  return this.frames(a, b, c);
+};
+
+Cut.Anim.prototype.frames = function(frames) {
   this._time = this._time || 0;
 
   this._frame = 0;
@@ -1097,7 +1145,7 @@ Cut.Anim.prototype.repeat = function(repeat, callback) {
 };
 
 Cut.Anim.prototype.play = function(frame) {
-  if (arguments.length) {
+  if (typeof frame !== 'undefined') {
     this.gotoFrame(frame);
     this._time = Cut._now();
   } else if (!this._time) {
@@ -1108,14 +1156,14 @@ Cut.Anim.prototype.play = function(frame) {
 
 Cut.Anim.prototype.stop = function(frame) {
   this._time = null;
-  if (arguments.length) {
+  if (typeof frame !== 'undefined') {
     this.gotoFrame(frame);
   }
   return this;
 };
 
-Cut.string = function(font) {
-  return new Cut.String().setFont(font);
+Cut.string = function(frames) {
+  return new Cut.String().frames(frames);
 };
 
 Cut.String = function() {
@@ -1127,21 +1175,35 @@ Cut.String.prototype = Cut._create(Cut.prototype);
 Cut.String.prototype._super = Cut;
 Cut.String.prototype.constructor = Cut.String;
 
-Cut.String.prototype.setFont = function(font) {
-  if (typeof font == 'string') {
-    this._font = function(value) {
-      return font + value;
+/**
+ * @deprecated Use frames
+ */
+Cut.String.prototype.setFont = function(a, b, c) {
+  return this.frames(a, b, c);
+};
+
+Cut.String.prototype.frames = function(frames) {
+  if (typeof frames == 'string') {
+    this._frames = function(value) {
+      return frames + value;
     };
-  } else if (Cut._isFunc(font)) {
-    this._font = font;
+  } else if (typeof frames === 'function') {
+    this._frames = frames;
   }
   return this;
 };
 
-Cut.String.prototype.setValue = function(value) {
-  if (this.value === value)
+/**
+ * @deprecated Use value
+ */
+Cut.String.prototype.setValue = function(a, b, c) {
+  return this.value(a, b, c);
+};
+
+Cut.String.prototype.value = function(value) {
+  if (this._value === value)
     return this;
-  this.value = value;
+  this._value = value;
 
   if (typeof value !== 'string' && !Cut._isArray(value)) {
     value = value + '';
@@ -1149,9 +1211,9 @@ Cut.String.prototype.setValue = function(value) {
 
   var child = this._first;
   for (var i = 0; i < value.length; i++) {
-    var selector = this._font(value[i]);
+    var selector = this._frames(value[i]);
     if (child) {
-      child.setImage(selector).show();
+      child.image(selector).show();
     } else {
       child = Cut.image(selector).appendTo(this);
     }
@@ -1515,7 +1577,12 @@ Cut.addTexture = function() {
 Cut.Out = function(def, image, ratio) {
   this.isCutout = true;
 
-  this._image = image;
+  if (typeof image === 'function') {
+    this._imagefn = image;
+  } else {
+    this._image = image;
+  }
+
   this._ratio = ratio || 1;
 
   this._name = def.name;
@@ -1557,7 +1624,7 @@ Cut.Out.prototype.clone = function() {
 };
 
 Cut.Out.prototype.dWidth = function(width) {
-  if (arguments.length) {
+  if (typeof width !== 'undefined') {
     this._dw = width;
     return this;
   }
@@ -1565,7 +1632,7 @@ Cut.Out.prototype.dWidth = function(width) {
 };
 
 Cut.Out.prototype.dHeight = function(height) {
-  if (arguments.length) {
+  if (typeof height !== 'undefined') {
     this._dh = height;
     return this;
   }
@@ -1598,12 +1665,16 @@ Cut.Out.prototype.offset = function(x, y) {
 
 Cut.Out.prototype.paste = function(context) {
   Cut._stats.paste++;
-  var img = typeof this._image === 'function' ? this._image() : this._image;
+  if (!this._image && this._imagefn) {
+    this._image = this._imagefn();
+  }
   try {
-    img && context.drawImage(img, // source
-    this._sx, this._sy, this._sw, this._sh, // cut
-    this._dx, this._dy, this._dw, this._dh // position
-    );
+    if (this._image) {
+      context.drawImage(this._image, // source
+      this._sx, this._sy, this._sw, this._sh, // cut
+      this._dx, this._dy, this._dw, this._dh // position
+      );
+    }
   } catch (e) {
     if (!this._failed) {
       console.log('Unable to paste: ' + this, img);
@@ -1863,46 +1934,29 @@ Cut.Pin.prototype.relativeMatrix = function() {
 
 Cut.Pin._SINGLE = {};
 
-Cut.Pin.prototype.update = function() {
+Cut.Pin.prototype.get = function(a) {
+  return Cut.Pin._getters[a] ? Cut.Pin._getters[a](this) : undefined;
+};
 
-  if (arguments.length == 1 && typeof arguments[0] === 'string') {
-    return this['_' + arguments[0]];
+Cut.Pin.prototype.set = function(a, b) {
+
+  var pin = Cut.Pin._SINGLE, key = null;
+
+  for (key in pin)
+    delete pin[key];
+
+  if (typeof a === 'string') {
+    pin[a] = b;
+
+  } else if (typeof a === 'object') {
+    for (key in a)
+      pin[key] = a[key];
   }
 
-  this._transform_flag = false;
-  this._translate_flag = false;
-
-  var pin = null, key = null, value = null;
-
-  if (arguments.length == 2 && typeof arguments[0] === 'string') {
-    for (key in Cut.Pin._SINGLE) {
-      delete Cut.Pin._SINGLE[key];
+  for (key in pin) {
+    if (typeof pin[key] !== 'undefined' && Cut.Pin._setters[key]) {
+      Cut.Pin._setters[key](this, pin[key], pin);
     }
-    (pin = Cut.Pin._SINGLE)[arguments[0]] = arguments[1];
-
-  } else if (arguments.length == 1 && typeof arguments[0] === 'object') {
-    pin = arguments[0];
-  }
-
-  // TODO: map write only keys to read/write keys
-
-  if (pin) {
-    for (key in pin) {
-      if (typeof (value = pin[key]) !== 'undefined') {
-        if (setter = Cut.Pin._setters[key]) {
-          setter.call(Cut.Pin._setters, this, value, pin);
-        }
-      }
-    }
-  }
-
-  if (this._translate_flag) {
-    this._translate_flag = false;
-    this._ts_translate = Cut._TS++;
-  }
-  if (this._transform_flag) {
-    this._transform_flag = false;
-    this._ts_transform = Cut._TS++;
   }
 
   if (this._owner) {
@@ -1911,6 +1965,94 @@ Cut.Pin.prototype.update = function() {
   }
 
   return this;
+};
+
+Cut.Pin._getters = {
+  alpha : function(pin) {
+    return pin._alpha;
+  },
+
+  textureAlpha : function(pin) {
+    return pin._textureAlpha;
+  },
+
+  width : function(pin) {
+    return pin._width;
+  },
+
+  height : function(pin) {
+    return pin._height;
+  },
+
+  // scale : function(pin) {
+  // },
+
+  scaleX : function(pin) {
+    return pin._scaleX;
+  },
+
+  scaleY : function(pin) {
+    return pin._scaleY;
+  },
+
+  // skew : function(pin) {
+  // },
+
+  skewX : function(pin) {
+    return pin._skewX;
+  },
+
+  skewY : function(pin) {
+    return pin._skewY;
+  },
+
+  rotation : function(pin) {
+    return pin._rotation;
+  },
+
+  // pivot : function(pin) {
+  // },
+
+  pivotX : function(pin) {
+    return pin._pivotX;
+  },
+
+  pivotY : function(pin) {
+    return pin._pivotY;
+  },
+
+  // offset : function(pin) {
+  // },
+
+  offsetX : function(pin) {
+    return pin._offsetX;
+  },
+
+  offsetY : function(pin) {
+    return pin._offsetY;
+  },
+
+  // align : function(pin) {
+  // },
+
+  alignX : function(pin) {
+    return pin._alignX;
+  },
+
+  alignY : function(pin) {
+    return pin._alignY;
+  },
+
+  // handle : function(pin) {
+  // },
+
+  handleX : function(pin) {
+    return pin._handleX;
+  },
+
+  handleY : function(pin) {
+    return pin._handleY;
+  }
 };
 
 Cut.Pin._setters = {
@@ -1925,96 +2067,96 @@ Cut.Pin._setters = {
   width : function(pin, value, set) {
     pin._width_ = value;
     pin._width = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   height : function(pin, value, set) {
     pin._height_ = value;
     pin._height = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   scale : function(pin, value, set) {
     pin._scaleX = value;
     pin._scaleY = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   scaleX : function(pin, value, set) {
     pin._scaleX = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   scaleY : function(pin, value, set) {
     pin._scaleY = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   skew : function(pin, value, set) {
     pin._skewX = value;
     pin._skewY = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   skewX : function(pin, value, set) {
     pin._skewX = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   skewY : function(pin, value, set) {
     pin._skewY = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   rotation : function(pin, value, set) {
     pin._rotation = value;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   pivot : function(pin, value, set) {
     pin._pivotX = value;
     pin._pivotY = value;
     pin._pivoted = true;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   pivotX : function(pin, value, set) {
     pin._pivotX = value;
     pin._pivoted = true;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   pivotY : function(pin, value, set) {
     pin._pivotY = value;
     pin._pivoted = true;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   offset : function(pin, value, set) {
     pin._offsetX = value;
     pin._offsetY = value;
-    pin._translate_flag = true;
+    pin._ts_translate = Cut._TS++;
   },
 
   offsetX : function(pin, value, set) {
     pin._offsetX = value;
-    pin._translate_flag = true;
+    pin._ts_translate = Cut._TS++;
   },
 
   offsetY : function(pin, value, set) {
     pin._offsetY = value;
-    pin._translate_flag = true;
+    pin._ts_translate = Cut._TS++;
   },
 
   align : function(pin, value, set) {
-    this.alignX.apply(this, arguments);
-    this.alignY.apply(this, arguments);
+    this.alignX(pin, value, set);
+    this.alignY(pin, value, set);
   },
 
   alignX : function(pin, value, set) {
     pin._alignX = value;
     pin._aligned = true;
-    pin._translate_flag = true;
+    pin._ts_translate = Cut._TS++;
 
     this.handleX(pin, value, set);
   },
@@ -2022,7 +2164,7 @@ Cut.Pin._setters = {
   alignY : function(pin, value, set) {
     pin._alignY = value;
     pin._aligned = true;
-    pin._translate_flag = true;
+    pin._ts_translate = Cut._TS++;
 
     this.handleY(pin, value, set);
   },
@@ -2035,13 +2177,13 @@ Cut.Pin._setters = {
   handleX : function(pin, value, set) {
     pin._handleX = value;
     pin._handled = true;
-    pin._translate_flag = true;
+    pin._ts_translate = Cut._TS++;
   },
 
   handleY : function(pin, value, set) {
     pin._handleY = value;
     pin._handled = true;
-    pin._translate_flag = true;
+    pin._ts_translate = Cut._TS++;
   },
 
   resizeMode : function(pin, value, set) {
@@ -2064,7 +2206,7 @@ Cut.Pin._setters = {
     }
     pin._scaleX = value / pin._width_;
     pin._width = pin._width_;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   resizeHeight : function(pin, value, set, force) {
@@ -2073,7 +2215,7 @@ Cut.Pin._setters = {
     }
     pin._scaleY = value / pin._height_;
     pin._height = pin._height_;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   scaleMode : function(pin, value, set) {
@@ -2093,7 +2235,7 @@ Cut.Pin._setters = {
       return;
     }
     pin._scaleX = value / pin._width_;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   scaleHeight : function(pin, value, set, force) {
@@ -2101,7 +2243,7 @@ Cut.Pin._setters = {
       return;
     }
     pin._scaleY = value / pin._height_;
-    pin._transform_flag = true;
+    pin._ts_transform = Cut._TS++;
   },
 
   matrix : function(pin, value, set) {
@@ -2290,21 +2432,19 @@ Cut.Matrix.prototype.mapY = function(x, y) {
 Cut.Math = {};
 
 Cut.Math.random = function(min, max) {
-  if (arguments.length == 0) {
+  if (typeof min === 'undefined') {
     max = 1, min = 0;
-  } else if (arguments.length == 1) {
+  } else if (typeof max === 'undefined') {
     max = min, min = 0;
   }
-  if (min == max) {
-    return min;
-  }
-  return Math.random() * (max - min) + min;
+  return min == max ? min : Math.random() * (max - min) + min;
 };
 
 Cut.Math.rotate = function(num, min, max) {
-  if (arguments.length < 3) {
-    max = min || 0;
-    min = 0;
+  if (typeof min === 'undefined') {
+    max = 1, min = 0;
+  } else if (typeof max === 'undefined') {
+    max = min, min = 0;
   }
   if (max > min) {
     num = (num - min) % (max - min);
@@ -2319,8 +2459,6 @@ Cut.Math.length = function(x, y) {
   return Math.sqrt(x * x + y * y);
 };
 
-Cut._TS = 0;
-
 Cut._isCut = function(obj) {
   return obj instanceof Cut;
 };
@@ -2329,86 +2467,36 @@ Cut._isNum = function(x) {
   return typeof x === 'number';
 };
 
-Cut._isFunc = function(x) {
-  return typeof x === 'function';
-};
-
 Cut._isArray = ('isArray' in Array) ? Array.isArray : function(value) {
   return Object.prototype.toString.call(value) === '[object Array]';
 };
 
-Cut._extend = function(base, extension, attribs) {
-  if (attribs) {
-    for (var i = 0; i < attribs.length; i++) {
-      var attr = attribs[i];
-      base[attr] = extension[attr];
-    }
-  } else {
-    for ( var attr in extension) {
-      base[attr] = extension[attr];
+Cut._extend = function() {
+  var base = {};
+  for (var i = 0; i < arguments.length; i++) {
+    var obj = arguments[i];
+    for ( var name in obj) {
+      if (obj.hasOwnProperty(name)) {
+        base[name] = obj[name];
+      }
     }
   }
   return base;
 };
 
-Cut._now = (function() {
-  if (typeof performance !== 'undefined' && performance.now) {
-    return function() {
-      return performance.now();
-    };
-  } else if (Date.now) {
-    return function() {
-      return Date.now();
-    };
-  } else {
-    return function() {
-      return +new Date();
-    };
-  }
-})();
-
-Cut._function = function(value) {
-  return typeof value === 'function' ? value : function() {
-    return value;
+if (typeof performance !== 'undefined' && performance.now) {
+  Cut._now = function() {
+    return performance.now();
   };
-};
-
-Cut._options = function(options) {
-  options.get = function(name) {
-    return typeof this[name] == 'function' ? this[name]() : this[name];
+} else if (Date.now) {
+  Cut._now = function() {
+    return Date.now();
   };
-  options.extend = function(obj) {
-    obj = typeof obj === 'object' ? obj : {};
-    for ( var name in this) {
-      obj[name] = name in obj ? obj[name] : this[name];
-    }
-    return obj;
+} else {
+  Cut._now = function() {
+    return +new Date();
   };
-  options.mixin = function(obj) {
-    if (typeof obj === 'object') {
-      for ( var name in obj) {
-        this[name] = obj[name];
-      }
-    }
-    return this;
-  };
-  return options;
-};
-
-Cut._status = function(msg) {
-  if (!(Cut._statusbox)) {
-    var statusbox = Cut._statusbox = document.createElement('div');
-    statusbox.style.position = 'absolute';
-    statusbox.style.color = 'black';
-    statusbox.style.background = 'white';
-    statusbox.style.zIndex = 999;
-    statusbox.style.top = '5px';
-    statusbox.style.right = '5px';
-    statusbox.style.padding = '1px 5px';
-    document.body.appendChild(statusbox);
-  }
-  Cut._statusbox.innerHTML = msg;
-};
+}
 
 Cut.Easing = (function() {
 
@@ -2589,6 +2677,6 @@ Cut.Easing.add({
   }
 });
 
-if (typeof module !== 'undefined') {
+if (typeof module !== 'undefined' && module.exports) {
   module.exports = Cut;
 }