+++ gulpfile.js
@@ -1,74 +1,74 @@
 var fs = require('fs');
 var args = require('minimist')(process.argv.slice(2));
+var browserify = require('browserify');
+
 var gulp = require('gulp');
 var concat = require('gulp-concat');
 var uglify = require('gulp-uglify');
 var wrap = require('gulp-wrap');
-var bump = require('gulp-bump');
+var rename = require('gulp-rename');
 var mocha = require('gulp-mocha');
+var transform = require('vinyl-transform');
+var source = require('vinyl-source-stream');
+var buffer = require('vinyl-buffer');
 
-var paths = {
-  web : [ 'cut-core.js', 'cut-loader.web.js', 'cut-mouse.js' ],
-  corova : [ 'cut-core.js', 'cut-loader.corova.js', 'cut-mouse.js' ],
-  fc : [ 'cut-core.js', 'cut-loader.fc.js', 'cut-mouse.js' ]
-};
+var pkg = require('./package.json');
 
-gulp.task('default', [ 'mocha', 'dist' ]);
+gulp.task('default', [ 'build' ]);
 
-gulp.task('web', dist(paths.web, 'cut.web.js', 'cut.web.min.js'));
-gulp.task('cordova', dist(paths.web, 'cut.cordova.js', 'cut.cordova.min.js'));
-gulp.task('fc', dist(paths.web, 'cut.fc.js', 'cut.fc.min.js'));
-gulp.task('dist', [ 'web', 'cordova', 'fc' ]);
+gulp.task('build', [ 'web', 'cordova', 'fc' ]);
 
-gulp.task('bump', function() {
-  var task = gulp.src([ './bower.json', './package.json' ]);
-  task = task.pipe(bump(args.bump ? {
-    version : args.bump
-  } : {}));
-  task = task.pipe(gulp.dest('./'));
-  return task;
-});
+gulp.task('web', dist('cut.web'));
+gulp.task('cordova', dist('cut.cordova'));
+gulp.task('fc', dist('cut.fc'));
 
-gulp.task('mocha', function() {
+gulp.task('web-nomin', dist('cut.web', true));
+gulp.task('cordova-nomin', dist('cut.cordova', true));
+gulp.task('fc-nomin', dist('cut.fc', true));
+gulp.task('build-nomin', [ 'web-nomin', 'cordova-nomin', 'fc-nomin' ]);
+
+gulp.task('test', function() {
   return gulp.src('test/*.js', {
     read : false
   }).pipe(mocha({}));
 });
 
-function dist(files, src, min) {
+gulp.task('watch', function() {
+  gulp.watch('{cut-*.js,temp/*}', [ 'test', 'build-nomin' ]);
+});
+
+function dist(file, nomin) {
   return function() {
-    var pkg = getPackageJson();
-    var task = gulp.src(files);
-    task = task.pipe(concat(src));
-    task = task
-        .pipe(uglify({
-          compress : false,
-          mangle : false,
-          output : {
-            beautify : true,
-            comments : function(node, comment) {
-              return comment.type != 'comment2'
-                  || !/@license/i.test(comment.value);
-            }
-          }
-        }));
+    var task = browserify({
+      entries : [ './temp/' + file ],
+      standalone : 'Cut'
+    });
+    task = task.transform({
+      fromString : true,
+      compress : false,
+      mangle : false,
+      output : {
+        beautify : true,
+        comments : /^((?!@license)[\s\S])*$/i
+      }
+    }, 'uglifyify');
+    task = task.bundle();
+    task = task.pipe(source(file + '.js')).pipe(buffer()); // vinylify
     task = task.pipe(wrap({
-      src : 'dist.js'
+      src : './temp/dist.js'
     }, {
       version : pkg.version
     }));
     task = task.pipe(gulp.dest('dist'));
-    task = task.pipe(concat(min));
-    task = task.pipe(uglify({
-      output : {
-        comments : /@license/i
-      }
-    }));
-    task = task.pipe(gulp.dest('dist'));
+    if (!nomin) {
+      task = task.pipe(rename(file + '.min.js'));
+      task = task.pipe(uglify({
+        output : {
+          comments : /@license/i
+        }
+      }));
+      task = task.pipe(gulp.dest('dist'));
+    }
     return task;
   };
-}
-
-function getPackageJson() {
-  return JSON.parse(fs.readFileSync('./package.json', 'utf8'));
 }